refactor: Change the terminology from 'container' to 'sandbox'
After implementing the use of tmpfs with Docker, it became clear that CreateContainer might not be equivalent to creating a container at the Docker level. In this context, CreateContainer not only creates a Docker container but also starts it. Meanwhile, StartContainer was actually responsible for executing a specific command. To resolve this ambiguity, it was decided to distinguish between the concepts of a Sandbox (a software abstraction over Docker containers) and a container, thereby eliminating confusion.